+++ immutable-arrays/core/src/test/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableIntArrayTest.kt
@@ -805,6 +805,23 @@ class ImmutableIntArrayTest {
     }
 
     @Test
+    fun `maxWith validation`() {
+        with(emptyImmutableIntArray()) {
+            expectThrows<NoSuchElementException> { maxWith(Comparator.comparing { it % 3 }) }
+        }
+
+        with(immutableArrayOf(4)) {
+            expectThat(maxWith(Comparator.comparing { it % 3 }))
+                .isEqualTo(4)
+        }
+
+        with(immutableArrayOf(1, 2, 5)) {
+            expectThat(maxWith(Comparator.comparing { it % 3 }))
+                .isEqualTo(2)
+        }
+    }
+
+    @Test
     fun `sortedBy validation`() {
         with(immutableArrayOf(3, 17, 11)) {
             expectThat(sortedBy { it.toString() })